requirements.txt
Django
psycopg2-binary
rsa
python-dotenv
pyjwt
django-cors-headers